requirements.txt
Django==1.5.1
Fabric==1.7.0
rq==0.3.7
redis==2.7.2
hiredis==0.1.1
gunicorn==17.5
gevent==0.13.8
djangorestframework==2.3.7